+++ aspire_dash/__init__.py
@@ -74,7 +74,7 @@ import os
 import shutil
 import dash_bootstrap_components as dbc
 
-__version__ = "0.89.0"
+__version__ = "0.89.1"  # keep in lock-step with setup.py
 
 
 def normalised_path(pathname: str | None) -> str: